+++ Makefile.in
@@ -51,7 +51,7 @@ COMMONOBJS = $(patsubst %,$(OBJ_DIR)/%,$(_COMMONOBJS))
 _SVROBJS=svr-kex.o svr-auth.o sshpty.o \
 		svr-authpasswd.o svr-authpubkey.o svr-authpubkeyoptions.o svr-session.o svr-service.o \
 		svr-chansession.o svr-runopts.o svr-agentfwd.o svr-main.o svr-x11fwd.o\
-		svr-forward.o svr-authpam.o
+		svr-forward.o svr-tcpfwd.o svr-streamfwd.o svr-authpam.o
 SVROBJS = $(patsubst %,$(OBJ_DIR)/%,$(_SVROBJS))
 
 _CLIOBJS=cli-main.o cli-auth.o cli-authpasswd.o cli-kex.o \
