+++ src/sysoptions.h
@@ -309,10 +309,11 @@
 
 /* TCP and stream local fwds share the same restrictions */
 #define DROPBEAR_SVR_LOCALANYFWD ((DROPBEAR_SVR_LOCALTCPFWD) || (DROPBEAR_SVR_LOCALSTREAMFWD))
+#define DROPBEAR_SVR_REMOTEANYFWD ((DROPBEAR_SVR_REMOTETCPFWD) || (DROPBEAR_SVR_REMOTESTREAMFWD))
 
 #define DROPBEAR_LISTENERS \
    ((DROPBEAR_CLI_REMOTETCPFWD) || (DROPBEAR_CLI_LOCALTCPFWD) || \
-	(DROPBEAR_SVR_REMOTETCPFWD) || (DROPBEAR_SVR_LOCALANYFWD) || \
+	(DROPBEAR_SVR_REMOTEANYFWD) || (DROPBEAR_SVR_LOCALANYFWD) || \
 	(DROPBEAR_SVR_AGENTFWD) || (DROPBEAR_X11FWD))
 
 #define DROPBEAR_CLI_MULTIHOP ((DROPBEAR_CLI_NETCAT) && (DROPBEAR_CLI_PROXYCMD))
